Added comparison operations for RealAlgebraicNumbers.
The operators are not fully implemented yet, as refinement methods for
RealAlgebraicNumbersIR are still missing.

Implementing them as functions (and relying on overloading) did not work, as
we need to use them with std::lower_bound() which fails to resolve it
template parameters in this case. Hence, they are implemented as functors.