Added new methods for the FormulaPool that allow for a meaningful Constraint normalization.
We already distinguish between a Formula (everything that is not negated) and a negated Formula (NOT Formula).

isBaseFormula(): Checks whether the given Constraint is the "base" constraint and not the "negated" constraint.

getBaseFormula(): Gets the Formula from a Formula or a negated Formula.

createNegatedFormula(): Creates a negated constraint instead of a Negation around a constraint.